pyproject.toml
[project]
name = "vgv-rag"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "supabase>=2.0.0",
    "pydantic-settings>=2.0.0",
    "notion-client>=2.0.0",
    "slack-sdk>=3.0.0",
    "PyGithub>=2.0.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.1",
    "pdfminer-six>=20260107",
    "voyageai>=0.3.7",
    "pinecone>=8.1.2",
    "pyjwt>=2.12.1",
    "cryptography>=46.0.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vgv_rag"]
